Moved LabelEditor to widgets
