+++ src/app/graph2/components/layers/GraphEdgeLayer.js
