+++ src/app/graph2/components/layers/GraphHighlightLayer.js
