+++ src/app/graph2/components/layers/GraphNodeLayer.js
