+++ src/app/graph2/components/layers/IndexedGraphStartMarkerLayer.js
