+++ src/app/graph2/components/layers/SelectionBoxLayer.js
