+++ src/app/graph2/components/layers/ViewportNavigationLayer.css
