+++ src/app/graph2/components/widgets/LabelEditorWidget.css
